Added License as an entity

+++ Backend/src/main/java/gruppe01/ntnu/no/Warehouse/Workflow/Assigner/DatabaseInit.sql
@@ -3,7 +3,7 @@ USE warehouse;
 
 CREATE TABLE IF NOT EXISTS worker
 (
-    id            INT AUTO_INCREMENT PRIMARY KEY,
+    id            BIGINT AUTO_INCREMENT PRIMARY KEY,
     name          VARCHAR(255) NOT NULL,
     work_title    VARCHAR(255) NOT NULL,
     effectiveness DOUBLE       NOT NULL
@@ -11,21 +11,21 @@ CREATE TABLE IF NOT EXISTS worker
 
 CREATE TABLE IF NOT EXISTS license
 (
-    id   INT AUTO_INCREMENT PRIMARY KEY,
+    id   BIGINT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS worker_license
 (
-    worker_id  INT NOT NULL,
-    license_id INT NOT NULL,
+    worker_id  BIGINT NOT NULL,
+    license_id BIGINT NOT NULL,
     FOREIGN KEY (worker_id) REFERENCES worker (id),
     FOREIGN KEY (license_id) REFERENCES license (id)
 );
 
 CREATE TABLE IF NOT EXISTS zone
 (
-    id          INT AUTO_INCREMENT PRIMARY KEY,
+    id          BIGINT AUTO_INCREMENT PRIMARY KEY,
     name        VARCHAR(255) NOT NULL,
     description TEXT,
     capacity    INT          NOT NULL
@@ -33,7 +33,7 @@ CREATE TABLE IF NOT EXISTS zone
 
 CREATE TABLE IF NOT EXISTS task
 (
-    id           INT AUTO_INCREMENT PRIMARY KEY,
+    id           BIGINT AUTO_INCREMENT PRIMARY KEY,
     name         VARCHAR(255) NOT NULL,
     description  TEXT,
     min_duration INT          NOT NULL,
@@ -44,8 +44,8 @@ CREATE TABLE IF NOT EXISTS task
 
 CREATE TABLE IF NOT EXISTS active_task
 (
-    id         INT AUTO_INCREMENT PRIMARY KEY,
-    task_id    INT  NOT NULL,
+    id         BIGINT AUTO_INCREMENT PRIMARY KEY,
+    task_id    BIGINT NOT NULL,
     due_date   TIMESTAMP NOT NULL,
     start_time TIMESTAMP NOT NULL,
     end_time   TIMESTAMP NOT NULL,
@@ -54,32 +54,32 @@ CREATE TABLE IF NOT EXISTS active_task
 
 CREATE TABLE IF NOT EXISTS active_task_worker
 (
-    active_task_id INT NOT NULL,
-    worker_id      INT NOT NULL,
+    active_task_id BIGINT NOT NULL,
+    worker_id      BIGINT NOT NULL,
     FOREIGN KEY (active_task_id) REFERENCES active_task (id),
     FOREIGN KEY (worker_id) REFERENCES worker (id)
 );
 
 CREATE TABLE IF NOT EXISTS task_license
 (
-    task_id    INT NOT NULL,
-    license_id INT NOT NULL,
+    task_id    BIGINT NOT NULL,
+    license_id BIGINT NOT NULL,
     FOREIGN KEY (task_id) REFERENCES task (id),
     FOREIGN KEY (license_id) REFERENCES license (id)
 );
 
 CREATE TABLE IF NOT EXISTS zone_task
 (
-    zone_id INT NOT NULL,
-    task_id INT NOT NULL,
+    zone_id BIGINT NOT NULL,
+    task_id BIGINT NOT NULL,
     FOREIGN KEY (zone_id) REFERENCES zone (id),
     FOREIGN KEY (task_id) REFERENCES task (id)
 );
 
 CREATE TABLE IF NOT EXISTS zone_worker
 (
-    zone_id   INT NOT NULL,
-    worker_id INT NOT NULL,
+    zone_id   BIGINT NOT NULL,
+    worker_id BIGINT NOT NULL,
     FOREIGN KEY (zone_id) REFERENCES zone (id),
     FOREIGN KEY (worker_id) REFERENCES worker (id)
 );
